refactor: refactor applink name
